requirements.txt
annotated-types==0.7.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pydantic==2.9.2
pydantic_core==2.23.4
PyYAML==6.0.2
typing_extensions==4.12.2
Werkzeug==3.0.4
